refactor: more memory collection cleanup
